Remove need for changes to SSH marshalling routines
Prior to this change, marshalling and unmarshalling messages was
dependent on a small change to the gosshnew SSH message marshalling
routines to support substructs. Now, we handle the small edge case
internally where such a substruct would have been useful.

Also, run gofmt, and use the sftp-server umask option instead of setting
it internally in the test.